fix: 세션 로그인 재구현 전 갈아엎기
- 이런 저런 정보를 찾고, 피드백을 받아본 결과, Spring Security와 Spring Session을 같이 쓰는 형태의 장점을 찾지 못함.
- 특히, 분산 환경을 고려해야하는 추후 구현에서, 오히려 컨텍스트 홀더에 의존적으로 개발을 하는 것은 올바른 선택이 아니라고 판단.
- 따라서 아예 스프링 시큐리티를 이번 프로젝트에선 배제하기로 결정, 대신 암호화 관련 라이브러리는 security-crypto라는 특정 암호화 라이브러리만 가져와서 사용
