--- src/main/java/com/reboot_course/firstcome_system/auth/security/CustomAuthenticationProvider.java
@@ -1,43 +1,0 @@
-package com.reboot_course.firstcome_system.auth.security;
-
-import com.reboot_course.firstcome_system.member.entity.Member;
-import com.reboot_course.firstcome_system.member.repository.MemberRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.BadCredentialsException;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Component;
-
-import java.util.Collections;
-
-@Component
-@RequiredArgsConstructor
-public class CustomAuthenticationProvider implements AuthenticationProvider {
-
-    private final MemberRepository memberRepository;
-    private final PasswordEncoder passwordEncoder;
-
-    @Override
-    public Authentication authenticate(Authentication authentication) throws AuthenticationException {
-        String email = authentication.getName();
-        String password = authentication.getCredentials().toString();
-
-        Member member = memberRepository.findByEmail(email)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found: " + email));
-
-        if (!passwordEncoder.matches(password, member.getPassword())) {
-            throw new BadCredentialsException("Invalid password");
-        }
-
-        return new UsernamePasswordAuthenticationToken(email, password, Collections.emptyList());
-    }
-
-    @Override
-    public boolean supports(Class<?> authentication) {
-        return authentication.equals(UsernamePasswordAuthenticationToken.class);
-    }
-}
