--- src/main/java/com/reboot_course/firstcome_system/auth/security/UserDetailsServiceImpl.java
@@ -1,28 +1,0 @@
-package com.reboot_course.firstcome_system.auth.security;
-
-import com.reboot_course.firstcome_system.member.entity.Member;
-import com.reboot_course.firstcome_system.member.repository.MemberRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class UserDetailsServiceImpl implements UserDetailsService {
-
-    private final MemberRepository memberRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        Member member = memberRepository.findByEmail(username)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
-
-        return org.springframework.security.core.userdetails.User
-                .withUsername(member.getEmail())
-                .password(member.getPassword())
-                .roles("USER")
-                .build();
-    }
-}
