+++ src/main/java/com/reboot_course/firstcome_system/auth/session/controller/SessionController.java
@@ -1,0 +1,33 @@
+package com.reboot_course.firstcome_system.auth.session.controller;
+
+import com.reboot_course.firstcome_system.auth.session.dto.request.LoginRequest;
+import com.reboot_course.firstcome_system.auth.session.service.AuthService;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Slf4j
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/session")
+public class SessionController {
+    private final AuthService authService;
+
+    @PostMapping("/login")
+    public ResponseEntity<String> login(@RequestBody LoginRequest loginRequest, HttpServletRequest request, HttpServletResponse response) {
+        String sessionId = authService.login(loginRequest, request);
+        return ResponseEntity.ok(sessionId);
+    }
+
+    @PostMapping("/logout")
+    public ResponseEntity<Void> logout(HttpServletRequest request, HttpServletResponse response) {
+        authService.logout(request);
+        return ResponseEntity.ok().build();
+    }
+}
