+++ src/main/java/com/reboot_course/firstcome_system/member/entity/Member.java
@@ -1,7 +1,7 @@
 package com.reboot_course.firstcome_system.member.entity;
 
-import com.reboot_course.firstcome_system.auth.converter.EmailConverter;
-import com.reboot_course.firstcome_system.auth.converter.PersonalInfoConverter;
+import com.reboot_course.firstcome_system.auth.encryption.converter.EmailConverter;
+import com.reboot_course.firstcome_system.auth.encryption.converter.PersonalInfoConverter;
 import com.reboot_course.firstcome_system.common.entity.BaseEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
